Introduce ErrorDigest
